refactor: #20 - Change module name to metacontroller.io
