+++ Dockerfile
@@ -2,12 +2,12 @@ FROM golang:1.10 AS build
 
 RUN curl https://raw.githubusercontent.com/golang/dep/master/install.sh | sh
 
-COPY . /go/src/metacontroller.app/
-WORKDIR /go/src/metacontroller.app/
+COPY . /go/src/metacontroller.io/
+WORKDIR /go/src/metacontroller.io/
 ENV CGO_ENABLED=0
 RUN dep ensure && go install
 
 FROM alpine:3.12.0
-COPY --from=build /go/bin/metacontroller.app /usr/bin/metacontroller
+COPY --from=build /go/bin/metacontroller.io /usr/bin/metacontroller
 RUN apk update && apk add --no-cache ca-certificates
 CMD ["/usr/bin/metacontroller"]
