+++ Dockerfile.dev
@@ -2,11 +2,11 @@
 # It assumes you already ran that locally.
 FROM golang:1.10 AS build
 
-COPY . /go/src/metacontroller.app/
-WORKDIR /go/src/metacontroller.app/
+COPY . /go/src/metacontroller.io/
+WORKDIR /go/src/metacontroller.io/
 RUN go install
 
 FROM debian:stretch-slim
 RUN apt-get update && apt-get install --no-install-recommends -y ca-certificates && rm -rf /var/lib/apt/lists/*
-COPY --from=build /go/bin/metacontroller.app /usr/bin/metacontroller
+COPY --from=build /go/bin/metacontroller.io /usr/bin/metacontroller
 CMD ["/usr/bin/metacontroller"]
