+++ Makefile
@@ -1,6 +1,6 @@
 TAG = dev
 
-PKG        := metacontroller.app
+PKG        := metacontroller.io
 API_GROUPS := metacontroller/v1alpha1
 
 all: install
@@ -18,10 +18,10 @@ integration-test:
 	PATH="$(PWD)/hack/bin:$(PATH)" go test ./test/integration/... -v -timeout 5m -args --logtostderr -v=5
 
 image: generated_files
-	docker build -t metacontroller/metacontroller:$(TAG) .
+	docker build -t metacontrollerio/metacontroller:$(TAG) .
 
 push: image
-	docker push metacontroller/metacontroller:$(TAG)
+	docker push metacontrollerio/metacontroller:$(TAG)
 
 # Code generators
 # https://github.com/kubernetes/community/blob/master/contributors/devel/api_changes.md#generate-code
