+++ controller/common/manage_children.go
@@ -27,10 +27,10 @@ import (
 	"k8s.io/apimachinery/pkg/util/diff"
 	utilerrors "k8s.io/apimachinery/pkg/util/errors"
 
-	"metacontroller.app/apis/metacontroller/v1alpha1"
-	dynamicapply "metacontroller.app/dynamic/apply"
-	dynamicclientset "metacontroller.app/dynamic/clientset"
-	k8s "metacontroller.app/third_party/kubernetes"
+	"metacontroller.io/apis/metacontroller/v1alpha1"
+	dynamicapply "metacontroller.io/dynamic/apply"
+	dynamicclientset "metacontroller.io/dynamic/clientset"
+	k8s "metacontroller.io/third_party/kubernetes"
 )
 
 func ApplyUpdate(orig, update *unstructured.Unstructured) (*unstructured.Unstructured, error) {
