+++ controller/composite/controller.go
@@ -33,16 +33,16 @@ import (
 	"k8s.io/client-go/tools/cache"
 	"k8s.io/client-go/util/workqueue"
 
-	"metacontroller.app/apis/metacontroller/v1alpha1"
-	mcclientset "metacontroller.app/client/generated/clientset/internalclientset"
-	mclisters "metacontroller.app/client/generated/lister/metacontroller/v1alpha1"
-	"metacontroller.app/controller/common"
-	"metacontroller.app/controller/common/finalizer"
-	dynamicclientset "metacontroller.app/dynamic/clientset"
-	dynamiccontrollerref "metacontroller.app/dynamic/controllerref"
-	dynamicdiscovery "metacontroller.app/dynamic/discovery"
-	dynamicinformer "metacontroller.app/dynamic/informer"
-	k8s "metacontroller.app/third_party/kubernetes"
+	"metacontroller.io/apis/metacontroller/v1alpha1"
+	mcclientset "metacontroller.io/client/generated/clientset/internalclientset"
+	mclisters "metacontroller.io/client/generated/lister/metacontroller/v1alpha1"
+	"metacontroller.io/controller/common"
+	"metacontroller.io/controller/common/finalizer"
+	dynamicclientset "metacontroller.io/dynamic/clientset"
+	dynamiccontrollerref "metacontroller.io/dynamic/controllerref"
+	dynamicdiscovery "metacontroller.io/dynamic/discovery"
+	dynamicinformer "metacontroller.io/dynamic/informer"
+	k8s "metacontroller.io/third_party/kubernetes"
 )
 
 type parentController struct {
@@ -119,7 +119,7 @@ func newParentController(resources *dynamicdiscovery.ResourceMap, dynClient *dyn
 		updateStrategy: updateStrategy,
 		queue:          workqueue.NewNamedRateLimitingQueue(workqueue.DefaultControllerRateLimiter(), "CompositeController-"+cc.Name),
 		finalizer: &finalizer.Manager{
-			Name:    "metacontroller.app/compositecontroller-" + cc.Name,
+			Name:    "metacontroller.io/compositecontroller-" + cc.Name,
 			Enabled: cc.Spec.Hooks.Finalize != nil,
 		},
 	}
