+++ controller/composite/controller_revision.go
@@ -25,7 +25,7 @@ import (
 	"sync"
 
 	"github.com/golang/glog"
-	"metacontroller.app/controller/common"
+	"metacontroller.io/controller/common"
 
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
@@ -33,9 +33,9 @@ import (
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/util/json"
 
-	"metacontroller.app/apis/metacontroller/v1alpha1"
-	dynamiccontrollerref "metacontroller.app/dynamic/controllerref"
-	k8s "metacontroller.app/third_party/kubernetes"
+	"metacontroller.io/apis/metacontroller/v1alpha1"
+	dynamiccontrollerref "metacontroller.io/dynamic/controllerref"
+	k8s "metacontroller.io/third_party/kubernetes"
 )
 
 const (
