+++ controller/decorator/metacontroller.go
@@ -28,14 +28,14 @@ import (
 	"k8s.io/client-go/tools/cache"
 	"k8s.io/client-go/util/workqueue"
 
-	"metacontroller.app/apis/metacontroller/v1alpha1"
-	mcinformers "metacontroller.app/client/generated/informer/externalversions"
-	mclisters "metacontroller.app/client/generated/lister/metacontroller/v1alpha1"
-	"metacontroller.app/controller/common"
-	dynamicclientset "metacontroller.app/dynamic/clientset"
-	dynamicdiscovery "metacontroller.app/dynamic/discovery"
-	dynamicinformer "metacontroller.app/dynamic/informer"
-	k8s "metacontroller.app/third_party/kubernetes"
+	"metacontroller.io/apis/metacontroller/v1alpha1"
+	mcinformers "metacontroller.io/client/generated/informer/externalversions"
+	mclisters "metacontroller.io/client/generated/lister/metacontroller/v1alpha1"
+	"metacontroller.io/controller/common"
+	dynamicclientset "metacontroller.io/dynamic/clientset"
+	dynamicdiscovery "metacontroller.io/dynamic/discovery"
+	dynamicinformer "metacontroller.io/dynamic/informer"
+	k8s "metacontroller.io/third_party/kubernetes"
 )
 
 type Metacontroller struct {
