+++ dynamic/controllerref/controller_revision.go
@@ -28,9 +28,9 @@ import (
 	utilerrors "k8s.io/apimachinery/pkg/util/errors"
 	"k8s.io/kubernetes/pkg/util/pointer"
 
-	"metacontroller.app/apis/metacontroller/v1alpha1"
-	client "metacontroller.app/client/generated/clientset/internalclientset/typed/metacontroller/v1alpha1"
-	k8s "metacontroller.app/third_party/kubernetes"
+	"metacontroller.io/apis/metacontroller/v1alpha1"
+	client "metacontroller.io/client/generated/clientset/internalclientset/typed/metacontroller/v1alpha1"
+	k8s "metacontroller.io/third_party/kubernetes"
 )
 
 type ControllerRevisionManager struct {
