+++ dynamic/controllerref/unstructured.go
@@ -28,8 +28,8 @@ import (
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	utilerrors "k8s.io/apimachinery/pkg/util/errors"
 
-	dynamicclientset "metacontroller.app/dynamic/clientset"
-	k8s "metacontroller.app/third_party/kubernetes"
+	dynamicclientset "metacontroller.io/dynamic/clientset"
+	k8s "metacontroller.io/third_party/kubernetes"
 )
 
 type UnstructuredManager struct {
