+++ main.go
@@ -32,7 +32,7 @@ import (
 	"k8s.io/client-go/rest"
 	"k8s.io/client-go/tools/clientcmd"
 
-	"metacontroller.app/server"
+	"metacontroller.io/server"
 
 	_ "k8s.io/client-go/plugin/pkg/client/auth/oidc"
 )
